Fix partition_spec_id and sort_order_id types from Long to Integer (#196)
Iceberg v4 TrackedFile defines spec_id (141) and sort_order_id (140) as
IntegerType (i32). The kernel was writing these as i64 (Long), causing
Parquet type mismatch errors when Iceberg reads manifests since Parquet
does not support Long to Integer demotion.

Change partition_spec_id from i64 to i32 and sort_order_id from
Option<i64> to Option<i32> in ContentTreeNodeEntry, the builder,
the reader, and the write transform expressions.

+++ kernel/src/content_tree/reader.rs
@@ -144,8 +144,8 @@ fn visit_metadata_entry_at<'a>(
         .transpose()?;
 
     // Extract scalar fields
-    let partition_spec_id: i64 = getters[13].get(row_index, "partition_spec_id")?;
-    let sort_order_id: Option<i64> = getters[14].get_opt(row_index, "sort_order_id")?;
+    let partition_spec_id: i32 = getters[13].get(row_index, "partition_spec_id")?;
+    let sort_order_id: Option<i32> = getters[14].get_opt(row_index, "sort_order_id")?;
     let record_count: i64 = getters[15].get(row_index, "record_count")?;
     let file_size_in_bytes: Option<i64> = getters[16].get_opt(row_index, "file_size_in_bytes")?;
